server/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-google-genai>=0.0.2
gradio>=3.0.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pydantic>=2.0.0
yt-dlp
openai-whisper[cpu]